chore: create service for game group routers
