+++ backend/app/application/routers/game_router.py
@@ -2,82 +2,19 @@ import os
 import shutil
 import json
 
-from pathlib import Path
 from typing import Optional, List
 from fastapi import APIRouter, Depends, UploadFile, File, Form
 from sqlalchemy.orm import Session
 from backend.app.data.database import get_db
 from backend.app.application.security import require_role
-from backend.app.data.models.game_model import Game, GameTag, GamePlatform, GameChangelog, GamePhoto, GameVideo, GameFollow
-from backend.app.data.models.tag_platform_model import Tag, Platform
-from backend.app.application.schemas.game_schema import GameCreate
-from backend.app.application.services.email_service import send_game_update_to_subscribers
-
-
-STORAGE_BASE = Path(__file__).resolve().parents[4] / "backend" / "storage" / "games"
+from backend.app.application.services.game_service import get_game as svc_get_game, create_game as svc_create_game, delete_game as svc_delete_game, update_game as svc_update_game
 
 router = APIRouter(prefix="/games", tags=["Games"])
 
 
 @router.get("/")
 def get_game(db: Session = Depends(get_db)):
-    games = db.query(Game).all()
-
-    if not games:
-        return {"error": "No games found"}
-
-    result = []
-    for g in games:
-        tag_links = db.query(GameTag).filter(GameTag.game_id == g.id).all()
-        tags = []
-        for tl in tag_links:
-            tag = db.query(Tag).filter(Tag.id == tl.tag_id).first()
-            if tag:
-                tags.append(tag.name)
-
-        plat_links = db.query(GamePlatform).filter(GamePlatform.game_id == g.id).all()
-        platforms = []
-        for pl in plat_links:
-            platform = db.query(Platform).filter(Platform.id == pl.platform_id).first()
-            if platform:
-                platforms.append(platform.name)
-
-        photos = [p.file_path for p in db.query(GamePhoto).filter(GamePhoto.game_id == g.id).all()]
-        videos = [v.file_path for v in db.query(GameVideo).filter(GameVideo.game_id == g.id).all()]
-
-        subs = db.query(GameFollow).filter(GameFollow.game_id == g.id).all()
-        subscriber_ids = [s.user_id for s in subs] if subs else []
-        changelogs = [
-            {
-                "id": c.id,
-                "version": c.version,
-                "description": c.description,
-                "date": c.date.isoformat() if c.date else None,
-            }
-            for c in db.query(GameChangelog).filter(GameChangelog.game_id == g.id).all()
-        ]
-
-        result.append(
-            {
-                "id": g.id,
-                "title": g.title,
-                "description": g.description,
-                "download_link": g.download_link,
-                "cover_img_path": g.cover_img_path,
-                "start_date": g.start_date.isoformat() if g.start_date else None,
-                "release_date": g.release_date.isoformat() if g.release_date else None,
-                "repository_link": g.repository_link,
-                "type": g.type,
-                "tags": tags,
-                "platforms": platforms,
-                "photos": photos,
-                "videos": videos,
-                "changelogs": changelogs,
-                "subscribers": subscriber_ids,
-            }
-        )
-
-    return result
+    return svc_get_game(db)
 
 
 @router.post("/")
@@ -89,209 +26,12 @@ def create_game(
     db: Session = Depends(get_db),
     _user=Depends(require_role('developer')),
 ):
-    try:
-        body_data = json.loads(body)
-    except Exception:
-        body_data = {}
-
-    body = GameCreate.parse_obj(body_data)
-
-    game = Game(
-        title=body.title,
-        description=body.description,
-        download_link=body.download_link,
-        cover_img_path=None,
-        type=body.type,
-        start_date=body.start_date,
-        release_date=body.release_date,
-        repository_link=body.repository_link,
-    )
-
-    db.add(game)
-    db.commit()
-    db.refresh(game)
-
-    safe_name = body.title.replace(" ", "_").lower()
-    target_dir = STORAGE_BASE / safe_name
-    os.makedirs(target_dir, exist_ok=True)
-
-    def _save_upload_file(upload: UploadFile) -> Optional[str]:
-        if not upload:
-            return None
-        dest = target_dir / upload.filename
-        try:
-            with open(dest, "wb") as out:
-                shutil.copyfileobj(upload.file, out)
-            return f"/storage/games/{safe_name}/{upload.filename}"
-        except Exception:
-            return None
-
-
-    for name in body.tags or []:
-        tag = db.query(Tag).filter(Tag.name == name).first()
-        if not tag:
-            tag = Tag(name=name)
-            db.add(tag)
-            db.commit()
-            db.refresh(tag)
-
-        # avoid duplicate game-tag links
-        existing_gt = db.query(GameTag).filter(GameTag.game_id == game.id, GameTag.tag_id == tag.id).first()
-        if not existing_gt:
-            gt = GameTag(game_id=game.id, tag_id=tag.id)
-            db.add(gt)
-
-    for name in body.platforms or []:
-        plat = db.query(Platform).filter(Platform.name == name).first()
-        if not plat:
-            plat = Platform(name=name)
-            db.add(plat)
-            db.commit()
-            db.refresh(plat)
-
-        # avoid duplicate game-platform links
-        existing_gp = db.query(GamePlatform).filter(GamePlatform.game_id == game.id, GamePlatform.platform_id == plat.id).first()
-        if not existing_gp:
-            gp = GamePlatform(game_id=game.id, platform_id=plat.id)
-            db.add(gp)
-
-    if cover_img:
-        saved_cover = _save_upload_file(cover_img)
-        if saved_cover:
-            game.cover_img_path = saved_cover
-            db.commit()
-            db.refresh(game)
-
-    for file in photos or []:
-        saved = _save_upload_file(file)
-        if saved:
-            db.add(GamePhoto(game_id=game.id, file_path=saved))
-
-    for file in videos or []:
-        saved = _save_upload_file(file)
-        if saved:
-            db.add(GameVideo(game_id=game.id, file_path=saved))
-
-    for c in body.changelogs or []:
-        # avoid duplicate changelog versions for the same game; update if exists
-        existing_cl = db.query(GameChangelog).filter(GameChangelog.game_id == game.id, GameChangelog.version == c.version).first()
-        if existing_cl:
-            existing_cl.description = c.description
-            existing_cl.date = c.date
-        else:
-            changelog = GameChangelog(
-                game_id=game.id,
-                version=c.version,
-                description=c.description,
-                date=c.date,
-            )
-            db.add(changelog)
-
-    db.commit()
-
-    return {"id": game.id}
+    return svc_create_game(body, cover_img, photos, videos, db, _user)
 
 
 @router.delete("/{remove_id}")
 def delete_game(remove_id: int, db: Session = Depends(get_db), _user=Depends(require_role('developer'))):
-    game = db.query(Game).filter(Game.id == remove_id).first()
-
-    if not game:
-        return {"error": "Game not found"}
-
-    response = {"id": game.id, "name": game.title}
-
-    def _resolve(path_str: Optional[str]) -> Optional[Path]:
-        if not path_str:
-            return None
-        rel = path_str.replace("/storage/games/", "").lstrip("/")
-        return STORAGE_BASE / rel
-
-    try:
-        cover_path = _resolve(game.cover_img_path)
-        if cover_path and cover_path.exists():
-            cover_path.unlink()
-    except Exception:
-        pass
-
-    try:
-        download_link = _resolve(game.download_link)
-        if download_link and download_link.exists():
-            download_link.unlink()
-    except Exception:
-        pass
-
-    photos = db.query(GamePhoto).filter(GamePhoto.game_id == game.id).all()
-    for p in photos:
-        try:
-            p_path = _resolve(p.file_path)
-            if p_path and p_path.exists():
-                p_path.unlink()
-        except Exception:
-            pass
-        try:
-            db.delete(p)
-        except Exception:
-            pass
-
-    videos = db.query(GameVideo).filter(GameVideo.game_id == game.id).all()
-    for v in videos:
-        try:
-            v_path = _resolve(v.file_path)
-            if v_path and v_path.exists():
-                v_path.unlink()
-        except Exception:
-            pass
-        try:
-            db.delete(v)
-        except Exception:
-            pass
-
-    changelogs = db.query(GameChangelog).filter(GameChangelog.game_id == game.id).all()
-    for c in changelogs:
-        try:
-            db.delete(c)
-        except Exception:
-            pass
-
-    tag_links = db.query(GameTag).filter(GameTag.game_id == game.id).all()
-    for tl in tag_links:
-        try:
-            db.delete(tl)
-        except Exception:
-            pass
-
-    plat_links = db.query(GamePlatform).filter(GamePlatform.game_id == game.id).all()
-    for pl in plat_links:
-        try:
-            db.delete(pl)
-        except Exception:
-            pass
-
-    safe_name = None
-    candidates = [game.cover_img_path, game.download_link]
-    candidates += [p.file_path for p in photos]
-    candidates += [v.file_path for v in videos]
-    for c in candidates:
-        if c:
-            rel = c.replace("/storage/games/", "").lstrip("/")
-            parts = rel.split("/", 1)
-            if parts:
-                safe_name = parts[0]
-                break
-
-    if safe_name:
-        try:
-            dir_path = STORAGE_BASE / safe_name
-            if dir_path.exists():
-                shutil.rmtree(dir_path)
-        except Exception:
-            pass
-
-    db.delete(game)
-    db.commit()
-
-    return response
+    return svc_delete_game(remove_id, db, _user)
     
 
 @router.put("/{game_id}")
@@ -306,221 +46,4 @@ def update_game(
     db: Session = Depends(get_db),
     _user=Depends(require_role('developer')),
 ):
-    try:
-        body_data = json.loads(body)
-    except Exception:
-        body_data = {}
-
-    body = GameCreate.parse_obj(body_data)
-
-    game = db.query(Game).filter(Game.id == game_id).first()
-    game = db.query(Game).filter(Game.id == game_id).first()
-    if not game:
-        return {"error": "Game not found"}
-
-    old_title = game.title or ""
-
-    game.title = body.title
-    game.description = body.description
-    game.type = body.type
-    game.start_date = body.start_date
-    game.release_date = body.release_date
-    game.repository_link = body.repository_link
-
-    new_title = body.title or ""
-    old_safe = old_title.replace(" ", "_").lower() if old_title else None
-    new_safe = new_title.replace(" ", "_").lower()
-
-    def _replace_storage_path(path: Optional[str], old: str, new: str) -> Optional[str]:
-        if not path:
-            return None
-        return path.replace(f"/storage/games/{old}/", f"/storage/games/{new}/")
-
-    if old_title and old_title != new_title:
-        game.download_link = _replace_storage_path(game.download_link, old_safe, new_safe)
-        game.cover_img_path = _replace_storage_path(game.cover_img_path, old_safe, new_safe)
-
-        photos = db.query(GamePhoto).filter(GamePhoto.game_id == game.id).all()
-        for p in photos:
-            p.file_path = _replace_storage_path(p.file_path, old_safe, new_safe)
-
-        videos = db.query(GameVideo).filter(GameVideo.game_id == game.id).all()
-        for v in videos:
-            v.file_path = _replace_storage_path(v.file_path, old_safe, new_safe)
-
-        db.commit()
-        db.refresh(game)
-
-        old_dir = STORAGE_BASE / old_safe
-        new_dir = STORAGE_BASE / new_safe
-        try:
-            if old_dir.exists():
-                os.makedirs(new_dir.parent, exist_ok=True)
-                try:
-                    shutil.move(str(old_dir), str(new_dir))
-                except Exception:
-                    # fallback: move contents into new_dir
-                    os.makedirs(new_dir, exist_ok=True)
-                    for item in old_dir.iterdir():
-                        shutil.move(str(item), str(new_dir))
-                    try:
-                        old_dir.rmdir()
-                    except Exception:
-                        pass
-        except Exception:
-            pass
-
-    safe_name = new_safe
-    target_dir = STORAGE_BASE / safe_name
-
-    def _save_upload_file(upload: UploadFile) -> Optional[str]:
-        if not upload:
-            return None
-        dest = target_dir / upload.filename
-        try:
-            with open(dest, "wb") as out:
-                shutil.copyfileobj(upload.file, out)
-            return f"/storage/games/{safe_name}/{upload.filename}"
-        except Exception:
-            return None
-
-    def _resolve(path_str: Optional[str]) -> Optional[Path]:
-        if not path_str:
-            return None
-        rel = path_str.replace("/storage/games/", "").lstrip("/")
-        return STORAGE_BASE / rel
-
-    if photos_to_delete:
-        for item in photos_to_delete.split(","):
-            name = item.strip()
-            if not name:
-                continue
-            for p in db.query(GamePhoto).filter(GamePhoto.game_id == game.id).all():
-                try:
-                    if p.file_path == name or p.file_path.endswith(f"/{name}") or p.file_path.endswith(name):
-                        p_path = _resolve(p.file_path)
-                        if p_path and p_path.exists():
-                            p_path.unlink()
-                        db.delete(p)
-                except Exception:
-                    pass
-
-    if videos_to_delete:
-        for item in videos_to_delete.split(","):
-            name = item.strip()
-            if not name:
-                continue
-            for v in db.query(GameVideo).filter(GameVideo.game_id == game.id).all():
-                try:
-                    if v.file_path == name or v.file_path.endswith(f"/{name}") or v.file_path.endswith(name):
-                        v_path = _resolve(v.file_path)
-                        if v_path and v_path.exists():
-                            v_path.unlink()
-                        db.delete(v)
-                except Exception:
-                    pass
-
-    try:
-        new_tag_names = set(body.tags or [])
-        existing_gt_links = db.query(GameTag).filter(GameTag.game_id == game.id).all()
-        for gt in existing_gt_links:
-            tag = db.query(Tag).filter(Tag.id == gt.tag_id).first()
-            if tag and tag.name not in new_tag_names:
-                try:
-                    db.delete(gt)
-                except Exception:
-                    pass
-    except Exception:
-        pass
-
-    for name in body.tags or []:
-        tag = db.query(Tag).filter(Tag.name == name).first()
-        if not tag:
-            tag = Tag(name=name)
-            db.add(tag)
-            db.commit()
-            db.refresh(tag)
-
-        # avoid duplicate game-tag links
-        existing_gt = db.query(GameTag).filter(GameTag.game_id == game.id, GameTag.tag_id == tag.id).first()
-        if not existing_gt:
-            gt = GameTag(game_id=game.id, tag_id=tag.id)
-            db.add(gt)
-
-    for name in body.platforms or []:
-        plat = db.query(Platform).filter(Platform.name == name).first()
-        if not plat:
-            plat = Platform(name=name)
-            db.add(plat)
-            db.commit()
-            db.refresh(plat)
-
-        # avoid duplicate game-platform links
-        existing_gp = db.query(GamePlatform).filter(GamePlatform.game_id == game.id, GamePlatform.platform_id == plat.id).first()
-        if not existing_gp:
-            gp = GamePlatform(game_id=game.id, platform_id=plat.id)
-            db.add(gp)
-
-    try:
-        new_platform_names = set(body.platforms or [])
-        existing_gp_links = db.query(GamePlatform).filter(GamePlatform.game_id == game.id).all()
-        for gp in existing_gp_links:
-            plat = db.query(Platform).filter(Platform.id == gp.platform_id).first()
-            if plat and plat.name not in new_platform_names:
-                try:
-                    db.delete(gp)
-                except Exception:
-                    pass
-    except Exception:
-        pass
-
-    if cover_img:
-        if game.cover_img_path:
-            try:
-                rel = game.cover_img_path.replace("/storage/games/", "")
-                old_cover = STORAGE_BASE / rel
-                if old_cover.exists():
-                    old_cover.unlink()
-            except Exception:
-                pass
-
-        saved_cover = _save_upload_file(cover_img)
-        if saved_cover:
-            game.cover_img_path = saved_cover
-            db.commit()
-            db.refresh(game)
-
-    for file in photos or []:
-        saved = _save_upload_file(file)
-        if saved:
-            db.add(GamePhoto(game_id=game.id, file_path=saved))
-
-    for file in videos or []:
-        saved = _save_upload_file(file)
-        if saved:
-            db.add(GameVideo(game_id=game.id, file_path=saved))
-
-    for c in body.changelogs or []:
-        # avoid duplicate changelog versions for the same game; update if exists
-        existing_cl = db.query(GameChangelog).filter(GameChangelog.game_id == game.id, GameChangelog.version == c.version).first()
-        if existing_cl:
-            existing_cl.description = c.description
-            existing_cl.date = c.date
-        else:
-            changelog = GameChangelog(
-                game_id=game.id,
-                version=c.version,
-                description=c.description,
-                date=c.date,
-            )
-            db.add(changelog)
-
-            try:
-                send_game_update_to_subscribers(game.id, c.description or "", db)
-            except Exception:
-                pass
-
-
-    db.commit()
-
-    return {"id": game.id}
+    return svc_update_game(game_id, body, cover_img, photos, videos, photos_to_delete, videos_to_delete, db, _user)
